define int long long

+++ algo/disjoint-set.cpp
@@ -1,13 +1,13 @@
 /* Disjoint Set */
 class DisjointSet {
    private:
-    vector<ll> par, sizes;
+    vector<int> par, sizes;
 
    public:
-    DisjointSet(ll n) : par(n), sizes(n, 1) { iota(par.begin(), par.end(), 0); }
-    ll find(ll x) { return (par[x] == x ? x : par[x] = find(par[x])); }
-    bool unite(ll x, ll y) {
-        ll x_root = find(x), y_root = find(y);
+    DisjointSet(int n) : par(n), sizes(n, 1) { iota(par.begin(), par.end(), 0); }
+    int find(int x) { return (par[x] == x ? x : par[x] = find(par[x])); }
+    bool unite(int x, int y) {
+        int x_root = find(x), y_root = find(y);
         if (x_root == y_root)
             return false;
         if (sizes[x_root] < sizes[y_root])
@@ -16,5 +16,5 @@ class DisjointSet {
         par[y_root] = x_root;
         return true;
     }
-    ll tree_len(ll x) { return sizes[find(x)]; }
+    int tree_len(int x) { return sizes[find(x)]; }
 };
